+++ src/main/java/me/xhsun/guildwars2wrapper/GuildWars2.java
@@ -15,7 +15,8 @@ import retrofit2.converter.gson.GsonConverterFactory;
  * &#32;2) use methods provided by {@link AsynchronousRequest} to get/process data asynchronously
  * TODO chatLinkToItemID
  * TODO language selection, default english
- * //TODO method to resolve back story question using answer id
+ * TODO method to resolve back story question using answer id
+ * TODO coins to gold/sliver/copper
  * @author xhsun
  * @since 2017-02-06
  */
